#52 refactor: MemberErrorCode 변경에 따른 import 수정
